system1/agent/requirements.txt
openai
autogen
autogen-ext[mcp]
autogen-agentchat
fastmcp
a2a-sdk
httpx
